Renamed todo-list to add_todo. This will ONLY act as a way to add todo through our modal. Finished with our add todo.
