+++ README.md
@@ -11,7 +11,7 @@
 
 ## Objectives
 * [x] To-do List
-* [ ] Weather App
+* [x] Weather App
 * [x] News App
 * [x] Quote of the Day
 * [ ] Deploy with Netlify or maybe Firebase Hosting
@@ -21,4 +21,4 @@
 * [x] Tied Vue with Firebase realtime database
 * [x] Used axios to fetch and post data
 * [x] Tie the child component to Home.Vue
-* [ ] Refactor To-do List code. Is there a way to reduce the amount of GET request?
+* [x] Refactor To-do List code. Is there a way to reduce the amount of GET request?
